Internalize SqlBackend (#1225)
* abstractification

* abstractification

* Internalize SqlBackend

* suppor tother sql dbs

* bounds

* asdf

* import data.monoid

* limitoffset and backendspecificoverrides

* use setter

* getConnUpsertSql

* formatting

* sigh

* clean warns

* cabal formatting

* merge resolve

* lol

* ok

* update changelogs

* one more [ci skip]
