+++ persistent-postgresql/ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for persistent-postgresql
 
+
+## 2.13.0.0 (unreleased)
+
+* [#1225](https://github.com/yesodweb/persistent/pull/1225)
+    * Support `persistent-2.13.0.0` making SQlBackend internal
+
 # 2.12.1.1
 
 * [#1235](https://github.com/yesodweb/persistent/pull/1235)
