+++ persistent-sqlite/test/main.hs
@@ -176,160 +176,161 @@ setup migration = do
 
 main :: IO ()
 main = do
-  handle (\(_ :: IOException) -> return ())
-    $ removeFile $ fromText sqlite_database_file
+    handle (\(_ :: IOException) -> return ())
+        $ removeFile $ fromText sqlite_database_file
 
-  runConn $ do
-    mapM_ setup
-      [ ForeignKey.compositeMigrate
-      , PersistentTest.testMigrate
-      , PersistentTest.noPrefixMigrate
-      , PersistentTest.customPrefixMigrate
-      , EmbedTest.embedMigrate
-      , EmbedOrderTest.embedOrderMigrate
-      , LargeNumberTest.numberMigrate
-      , UniqueTest.uniqueMigrate
-      , MaxLenTest.maxlenMigrate
-      , Recursive.recursiveMigrate
-      , CompositeTest.compositeMigrate
-      , MigrationTest.migrationMigrate
-      , PersistUniqueTest.migration
-      , RenameTest.migration
-      , CustomPersistFieldTest.customFieldMigrate
-      , PrimaryTest.migration
-      , CustomPrimaryKeyReferenceTest.migration
-      , MigrationColumnLengthTest.migration
-      , TransactionLevelTest.migration
-      , LongIdentifierTest.migration
-      ]
-    PersistentTest.cleanDB
-    ForeignKey.cleanDB
+    runConn $ do
+        mapM_ setup
+            [ ForeignKey.compositeMigrate
+            , PersistentTest.testMigrate
+            , PersistentTest.noPrefixMigrate
+            , PersistentTest.customPrefixMigrate
+            , EmbedTest.embedMigrate
+            , EmbedOrderTest.embedOrderMigrate
+            , LargeNumberTest.numberMigrate
+            , UniqueTest.uniqueMigrate
+            , MaxLenTest.maxlenMigrate
+            , Recursive.recursiveMigrate
+            , CompositeTest.compositeMigrate
+            , MigrationTest.migrationMigrate
+            , PersistUniqueTest.migration
+            , RenameTest.migration
+            , CustomPersistFieldTest.customFieldMigrate
+            , PrimaryTest.migration
+            , CustomPrimaryKeyReferenceTest.migration
+            , MigrationColumnLengthTest.migration
+            , TransactionLevelTest.migration
+            , LongIdentifierTest.migration
+            ]
+        PersistentTest.cleanDB
+        ForeignKey.cleanDB
 
-  hspec $ do
-    RenameTest.specsWith db
-    DataTypeTest.specsWith
-        db
-        (Just (runMigrationSilent dataTypeMigrate))
-        [ TestFn "text" dataTypeTableText
-        , TestFn "textMaxLen" dataTypeTableTextMaxLen
-        , TestFn "bytes" dataTypeTableBytes
-        , TestFn "bytesTextTuple" dataTypeTableBytesTextTuple
-        , TestFn "bytesMaxLen" dataTypeTableBytesMaxLen
-        , TestFn "int" dataTypeTableInt
-        , TestFn "intList" dataTypeTableIntList
-        , TestFn "intMap" dataTypeTableIntMap
-        , TestFn "bool" dataTypeTableBool
-        , TestFn "day" dataTypeTableDay
-        , TestFn "time" (DataTypeTest.roundTime . dataTypeTableTime)
-        , TestFn "utc" (DataTypeTest.roundUTCTime . dataTypeTableUtc)
-        ]
-        [ ("pico", dataTypeTablePico) ]
-        dataTypeTableDouble
-    HtmlTest.specsWith
-        db
-        (Just (runMigrationSilent HtmlTest.htmlMigrate))
-    EmbedTest.specsWith db
-    EmbedOrderTest.specsWith db
-    LargeNumberTest.specsWith db
-    UniqueTest.specsWith db
-    MaxLenTest.specsWith db
-    Recursive.specsWith db
-    SumTypeTest.specsWith db (Just (runMigrationSilent SumTypeTest.sumTypeMigrate))
-    MigrationOnlyTest.specsWith db
-        (Just
-            $ runMigrationSilent MigrationOnlyTest.migrateAll1
-            >> runMigrationSilent MigrationOnlyTest.migrateAll2
-        )
-    PersistentTest.specsWith db
-    PersistentTest.filterOrSpecs db
-    ReadWriteTest.specsWith db
-    RawSqlTest.specsWith db
-    UpsertTest.specsWith
-        db
-        UpsertTest.Don'tUpdateNull
-        UpsertTest.UpsertPreserveOldKey
 
-    MpsNoPrefixTest.specsWith db
-    MpsCustomPrefixTest.specsWith db
-    EmptyEntityTest.specsWith db (Just (runMigrationSilent EmptyEntityTest.migration))
-    CompositeTest.specsWith db
-    PersistUniqueTest.specsWith db
-    PrimaryTest.specsWith db
-    CustomPersistFieldTest.specsWith db
-    CustomPrimaryKeyReferenceTest.specsWith db
-    MigrationColumnLengthTest.specsWith db
-    EquivalentTypeTest.specsWith db
-    ForeignKey.specsWith db
-    TransactionLevelTest.specsWith db
-    MigrationTest.specsWith db
-    LongIdentifierTest.specsWith db
-    GeneratedColumnTestSQL.specsWith db
+    hspec $ do
+        RenameTest.specsWith db
+        DataTypeTest.specsWith
+            db
+            (Just (runMigrationSilent dataTypeMigrate))
+            [ TestFn "text" dataTypeTableText
+            , TestFn "textMaxLen" dataTypeTableTextMaxLen
+            , TestFn "bytes" dataTypeTableBytes
+            , TestFn "bytesTextTuple" dataTypeTableBytesTextTuple
+            , TestFn "bytesMaxLen" dataTypeTableBytesMaxLen
+            , TestFn "int" dataTypeTableInt
+            , TestFn "intList" dataTypeTableIntList
+            , TestFn "intMap" dataTypeTableIntMap
+            , TestFn "bool" dataTypeTableBool
+            , TestFn "day" dataTypeTableDay
+            , TestFn "time" (DataTypeTest.roundTime . dataTypeTableTime)
+            , TestFn "utc" (DataTypeTest.roundUTCTime . dataTypeTableUtc)
+            ]
+            [ ("pico", dataTypeTablePico) ]
+            dataTypeTableDouble
+        HtmlTest.specsWith
+            db
+            (Just (runMigrationSilent HtmlTest.htmlMigrate))
+        EmbedTest.specsWith db
+        EmbedOrderTest.specsWith db
+        LargeNumberTest.specsWith db
+        UniqueTest.specsWith db
+        MaxLenTest.specsWith db
+        Recursive.specsWith db
+        SumTypeTest.specsWith db (Just (runMigrationSilent SumTypeTest.sumTypeMigrate))
+        MigrationOnlyTest.specsWith db
+            (Just
+                $ runMigrationSilent MigrationOnlyTest.migrateAll1
+                >> runMigrationSilent MigrationOnlyTest.migrateAll2
+            )
+        PersistentTest.specsWith db
+        PersistentTest.filterOrSpecs db
+        ReadWriteTest.specsWith db
+        RawSqlTest.specsWith db
+        UpsertTest.specsWith
+            db
+            UpsertTest.Don'tUpdateNull
+            UpsertTest.UpsertPreserveOldKey
 
-    it "issue #328" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        void $ runMigrationSilent migrateAll
-        insert_ . Test $ read "2014-11-30 05:15:25.123Z"
-        [Single x] <- rawSql "select strftime('%s%f',time) from test" []
-        liftIO $ x `shouldBe` Just ("141732452525.123" :: String)
-    it "issue #339" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        void $ runMigrationSilent migrateAll
-        now <- liftIO getCurrentTime
-        tid <- insert $ Test now
-        Just (Test now') <- get tid
-        liftIO $ now' `shouldBe` now
-    it "issue #564" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
-        hClose h
-        conn <- Sqlite.open (T.pack fp)
-        Sqlite.close conn
-        return ()
-    it "issue #527" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        void $ runMigrationSilent migrateAll
-        insertMany_ $ replicate 1000 (Test $ read "2014-11-30 05:15:25.123Z")
+        MpsNoPrefixTest.specsWith db
+        MpsCustomPrefixTest.specsWith db
+        EmptyEntityTest.specsWith db (Just (runMigrationSilent EmptyEntityTest.migration))
+        CompositeTest.specsWith db
+        PersistUniqueTest.specsWith db
+        PrimaryTest.specsWith db
+        CustomPersistFieldTest.specsWith db
+        CustomPrimaryKeyReferenceTest.specsWith db
+        MigrationColumnLengthTest.specsWith db
+        EquivalentTypeTest.specsWith db
+        ForeignKey.specsWith db
+        TransactionLevelTest.specsWith db
+        MigrationTest.specsWith db
+        LongIdentifierTest.specsWith db
+        GeneratedColumnTestSQL.specsWith db
 
-    it "properly migrates to a composite primary key (issue #669)" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        void $ runMigrationSilent compositeSetup
-        void $ runMigrationSilent compositeMigrateTest
-        pure ()
+        it "issue #328" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
+            void $ runMigrationSilent migrateAll
+            insert_ . Test $ read "2014-11-30 05:15:25.123Z"
+            [Single x] <- rawSql "select strftime('%s%f',time) from test" []
+            liftIO $ x `shouldBe` Just ("141732452525.123" :: String)
+        it "issue #339" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
+            void $ runMigrationSilent migrateAll
+            now <- liftIO getCurrentTime
+            tid <- insert $ Test now
+            Just (Test now') <- get tid
+            liftIO $ now' `shouldBe` now
+        it "issue #564" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
+            hClose h
+            conn <- Sqlite.open (T.pack fp)
+            Sqlite.close conn
+            return ()
+        it "issue #527" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
+            void $ runMigrationSilent migrateAll
+            insertMany_ $ replicate 1000 (Test $ read "2014-11-30 05:15:25.123Z")
 
-    it "test migrating sparse primary keys (issue #1184)" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
-        hClose h
-        let connInfo = Lens.set fkEnabled False $ mkSqliteConnectionInfo (T.pack fp)
-        runSqliteInfo connInfo $ do
-            void $ runMigrationSilent idSetup
-            forM_ (map toSqlKey [1,3]) $ \key -> do
-                insertKey key (Simple "foo")
-                insert (SimpleReference key "test")
-
-            validateForeignKeys
-
-        runSqliteInfo connInfo $ do
-            void $ runMigrationSilent idMigrateTest
-            validateForeignKeys
-
-    it "test migrating sparse composite primary keys (issue #1184)" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
-        hClose h
-        let connInfo = Lens.set fkEnabled False $ mkSqliteConnectionInfo (T.pack fp)
-
-        runSqliteInfo connInfo $ do
+        it "properly migrates to a composite primary key (issue #669)" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
             void $ runMigrationSilent compositeSetup
-            forM_ [(1,"foo"),(3,"bar")] $ \(intKey, strKey) -> do
-                let key = SimpleCompositeKey strKey intKey
-                insertKey key (SimpleComposite intKey strKey)
-                insert (SimpleCompositeReference intKey strKey "test")
-
-            validateForeignKeys
-
-        runSqliteInfo connInfo $ do
             void $ runMigrationSilent compositeMigrateTest
-            validateForeignKeys
+            pure ()
 
-    it "afterException" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
-        void $ runMigrationSilent testMigrate
-        let catcher :: forall m. Monad m => SomeException -> m ()
-            catcher _ = return ()
-        insert_ $ Person "A" 0 Nothing
-        insert_ (Person "A" 1 Nothing) `catch` catcher
-        insert_ $ Person "B" 0 Nothing
-        return ()
+        it "test migrating sparse primary keys (issue #1184)" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
+            hClose h
+            let connInfo = Lens.set fkEnabled False $ mkSqliteConnectionInfo (T.pack fp)
+            runSqliteInfo connInfo $ do
+                void $ runMigrationSilent idSetup
+                forM_ (map toSqlKey [1,3]) $ \key -> do
+                    insertKey key (Simple "foo")
+                    insert (SimpleReference key "test")
+
+                validateForeignKeys
+
+            runSqliteInfo connInfo $ do
+                void $ runMigrationSilent idMigrateTest
+                validateForeignKeys
+
+        it "test migrating sparse composite primary keys (issue #1184)" $ asIO $ withSystemTempFile "test564.sqlite3"$ \fp h -> do
+            hClose h
+            let connInfo = Lens.set fkEnabled False $ mkSqliteConnectionInfo (T.pack fp)
+
+            runSqliteInfo connInfo $ do
+                void $ runMigrationSilent compositeSetup
+                forM_ [(1,"foo"),(3,"bar")] $ \(intKey, strKey) -> do
+                    let key = SimpleCompositeKey strKey intKey
+                    insertKey key (SimpleComposite intKey strKey)
+                    insert (SimpleCompositeReference intKey strKey "test")
+
+                validateForeignKeys
+
+            runSqliteInfo connInfo $ do
+                void $ runMigrationSilent compositeMigrateTest
+                validateForeignKeys
+
+        it "afterException" $ asIO $ runSqliteInfo (mkSqliteConnectionInfo ":memory:") $ do
+            void $ runMigrationSilent testMigrate
+            let catcher :: forall m. Monad m => SomeException -> m ()
+                catcher _ = return ()
+            insert_ $ Person "A" 0 Nothing
+            insert_ (Person "A" 1 Nothing) `catch` catcher
+            insert_ $ Person "B" 0 Nothing
+            return ()
 
 validateForeignKeys
     :: (MonadResource m, MonadReader env m, BackendCompatible SqlBackend env)
