+++ persistent-test/persistent-test.cabal
@@ -1,5 +1,5 @@
 name:            persistent-test
-version:         2.12.0.0
+version:         2.13.0.0
 license:         MIT
 license-file:    LICENSE
 author:          Michael Snoyman <michael@snoyman.com>
@@ -60,7 +60,7 @@ library
 
     build-depends: 
         base                     >= 4.9       && < 5
-      , persistent               >= 2.12      && < 2.13
+      , persistent               >= 2.13      && < 2.14
       , aeson                    >= 1.0
       , blaze-html               >= 0.9
       , bytestring               >= 0.10
