+++ persistent/Database/Persist/Sql.hs
@@ -19,10 +19,15 @@ module Database.Persist.Sql
     , transactionSaveWithIsolation
     , transactionUndo
     , transactionUndoWithIsolation
-    , IsolationLevel (..)
     , getStmtConn
+    , mkColumns
+    , BackendSpecificOverrides
+    , emptyBackendSpecificOverrides
+    , getBackendSpecificForeignKeyName
+    , setBackendSpecificForeignKeyName
+    , defaultAttribute
       -- * Internal
-    , module Database.Persist.Sql.Internal
+    , IsolationLevel(..)
     , decorateSQLWithLimitOffset
     ) where
 
@@ -36,7 +41,7 @@ import Database.Persist.Sql.Migration
 import Database.Persist.Sql.Raw
 import Database.Persist.Sql.Run hiding (rawAcquireSqlConn, rawRunSqlPool)
 import Database.Persist.Sql.Types
-import Database.Persist.Sql.Types.Internal (IsolationLevel(..))
+import Database.Persist.Sql.Types.Internal (IsolationLevel(..), SqlBackend(..))
 
 import Database.Persist.Sql.Orphan.PersistQuery
 import Database.Persist.Sql.Orphan.PersistStore
