+++ persistent/Database/Persist/Sql/Migration.hs
@@ -31,6 +31,7 @@ import System.IO.Silently (hSilence)
 import GHC.Stack
 
 import Database.Persist.Sql.Types
+import Database.Persist.Sql.Types.Internal
 import Database.Persist.Sql.Raw
 import Database.Persist.Types
 import Database.Persist.Sql.Orphan.PersistStore()
