+++ persistent/Database/Persist/Sql/Orphan/PersistStore.hs
@@ -43,6 +43,7 @@ import Database.Persist.Class ()
 import Database.Persist.Sql.Class (PersistFieldSql)
 import Database.Persist.Sql.Raw
 import Database.Persist.Sql.Types
+import Database.Persist.Sql.Types.Internal
 import Database.Persist.Sql.Util (
     dbIdColumns, keyAndEntityColumnNames, parseEntityValues, entityColumnNames
   , updatePersistValue, mkUpdateText, commaSeparated, mkInsertValues)
