+++ persistent/Database/Persist/Sql/Raw.hs
@@ -16,6 +16,8 @@ import qualified Data.Text as T
 
 import Database.Persist
 import Database.Persist.Sql.Types
+import Database.Persist.Sql.Types.Internal
+import Database.Persist.SqlBackend.Internal
 import Database.Persist.Sql.Class
 
 rawQuery :: (MonadResource m, MonadReader env m, BackendCompatible SqlBackend env)
@@ -66,8 +68,8 @@ rawExecuteCount sql vals = do
     return res
 
 getStmt
-  :: (MonadIO m, BackendCompatible SqlBackend backend)
-  => Text -> ReaderT backend m Statement
+  :: (MonadIO m, MonadReader backend m, BackendCompatible SqlBackend backend)
+  => Text -> m Statement
 getStmt sql = do
     conn <- projectBackend `liftM` ask
     liftIO $ getStmtConn conn sql
