+++ persistent/Database/Persist/SqlBackend/Internal/IsolationLevel.hs
@@ -1,0 +1,19 @@
+module Database.Persist.SqlBackend.Internal.IsolationLevel where
+
+import Data.String (IsString(..))
+import Data.Monoid ((<>)) -- TODO: remove when GHC-8.2 support is dropped
+
+-- | Please refer to the documentation for the database in question for a full
+-- overview of the semantics of the varying isloation levels
+data IsolationLevel = ReadUncommitted
+                    | ReadCommitted
+                    | RepeatableRead
+                    | Serializable
+                    deriving (Show, Eq, Enum, Ord, Bounded)
+
+makeIsolationLevelStatement :: (Monoid s, IsString s) => IsolationLevel -> s
+makeIsolationLevelStatement l = "SET TRANSACTION ISOLATION LEVEL " <> case l of
+    ReadUncommitted -> "READ UNCOMMITTED"
+    ReadCommitted -> "READ COMMITTED"
+    RepeatableRead -> "REPEATABLE READ"
+    Serializable -> "SERIALIZABLE"
