+++ stack-8.10.yaml
@@ -1,0 +1,10 @@
+resolver: lts-17.8
+packages:
+  - ./persistent
+  - ./persistent-sqlite
+  - ./persistent-test
+  - ./persistent-mongoDB
+  - ./persistent-mysql
+  - ./persistent-postgresql
+  - ./persistent-redis
+  - ./persistent-qq
